BulkCopy Fix

+++ MS.Katusha.Services/UtilityService.cs
@@ -1,10 +1,7 @@
-﻿using System.Collections.Generic;
-using System.IO;
+﻿using System.IO;
 using MS.Katusha.Domain;
 using MS.Katusha.Infrastructure;
 using MS.Katusha.Interfaces.Services;
-using MS.Katusha.Repositories.DB;
-using Raven.Client;
 
 namespace MS.Katusha.Services
 {
@@ -20,9 +17,11 @@ namespace MS.Katusha.Services
         public void ClearDatabase(string photosFolder) {
             ReloadResources.ClearDatabase(_dbContext);
             RavenHelper.ClearRaven();
-            foreach(var fileName in Directory.EnumerateFiles(photosFolder)) {
-                File.Delete(fileName);
-            }
+            if (!Directory.Exists(photosFolder))
+                Directory.CreateDirectory(photosFolder);
+            else
+                foreach(var fileName in Directory.EnumerateFiles(photosFolder)) 
+                    File.Delete(fileName);
         }
 
         public void RegisterRaven()
